add orphanRemoval to Order class

+++ src/main/java/com/solstice/orderorderlines/service/OrderOrderLineService.java
@@ -111,7 +111,9 @@ public class OrderOrderLineService {
   public OrderLineItem deleteOrderLineItem(long orderId, long orderLineId) {
     OrderLineItem deletedOrderLineItem = orderLineItemRepository
         .findOrderLineItemByIdAndOrderId(orderLineId, orderId);
-    orderLineItemRepository.delete(deletedOrderLineItem);
+    if (deletedOrderLineItem != null) {
+      orderLineItemRepository.delete(deletedOrderLineItem);
+    }
     return deletedOrderLineItem;
   }
